requirements.txt
PyYAML
numpy
pandas
google-cloud-vision